Move getItemClassNames into ContextualMenu Deprecated tests
