+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenuItem.test.tsx
@@ -5,9 +5,8 @@ import { shallow, ShallowWrapper } from 'enzyme';
 import { ContextualMenuItemBase } from './ContextualMenuItem.base';
 import { IContextualMenuItemProps } from './ContextualMenuItem.types';
 import { IContextualMenuItem } from './ContextualMenu.types';
-import { IMenuItemClassNames, getItemClassNames } from './ContextualMenu.classNames';
+import { IMenuItemClassNames } from './ContextualMenu.classNames';
 import { hasSubmenu } from '../../utilities/contextualMenu/index';
-import { createTheme } from '../../Styling';
 
 describe('ContextMenuItemChildren', () => {
   describe('when a checkmark icon', () => {
@@ -101,65 +100,6 @@ describe('ContextMenuItemChildren', () => {
     it('renders the menu icon', () => {
       expect(wrapper).toMatchSnapshot();
     });
-  });
-});
-
-describe('getItemClassNames', () => {
-  // This test exists to validate that getItemClassNames signature has not changed
-  // to avoid breaking internal partners relying on this private API.
-  // See: https://github.com/OfficeDev/office-ui-fabric-react/pull/6738
-  it('accepts an argument list of style props values', () => {
-    const NoClassNamesTheme = createTheme({ disableGlobalClassNames: true });
-    const isDisabled = false;
-    const isExpanded = false;
-    const isChecked = false;
-    const isAnchorLink = false;
-    const isKnownIcon = true;
-    const dividerClassName = 'dividerFoo';
-    const itemClassName = 'foo';
-    const iconClassName = 'iconFoo';
-
-    const itemClassNames = getItemClassNames(
-      NoClassNamesTheme,
-      isDisabled,
-      isExpanded,
-      isChecked,
-      isAnchorLink,
-      isKnownIcon,
-      itemClassName,
-      dividerClassName,
-      iconClassName
-    );
-
-    expect(itemClassNames).toBeDefined();
-  });
-
-  it('applies custom classNames to style slots', () => {
-    const NoClassNamesTheme = createTheme({ disableGlobalClassNames: true });
-    const isDisabled = false;
-    const isExpanded = false;
-    const isChecked = false;
-    const isAnchorLink = false;
-    const isKnownIcon = true;
-    const dividerClassName = 'dividerFoo';
-    const itemClassName = 'foo';
-    const iconClassName = 'iconFoo';
-
-    const itemClassNames = getItemClassNames(
-      NoClassNamesTheme,
-      isDisabled,
-      isExpanded,
-      isChecked,
-      isAnchorLink,
-      isKnownIcon,
-      itemClassName,
-      dividerClassName,
-      iconClassName
-    );
-
-    expect(itemClassNames.item).toContain('foo');
-    expect(itemClassNames.divider).toContain('dividerFoo');
-    expect(itemClassNames.icon).toContain('iconFoo');
   });
 });
